moved the test main script out of the logic into a seperate main file
