+++ src/hvl/projectparmorel/ml/QLearning.java
@@ -69,7 +69,7 @@ public class QLearning {
 	public URI uri;
 	List<Error> original = new ArrayList<Error>();
 	List<Integer> processed = new ArrayList<Integer>();
-	List<Integer> originalCodes = new ArrayList<Integer>();
+	public List<Integer> originalCodes = new ArrayList<Integer>();
 	Map<Integer, double[]> Q = new HashMap<Integer, double[]>();
 	Map<Integer, Integer> errorMap = new HashMap<Integer, Integer>();
 	List<Sequence> solvingMap = new ArrayList<Sequence>();
@@ -84,7 +84,7 @@ public class QLearning {
 	public List<Error> nuQueue = new ArrayList<Error>();
 	NotificationChain msgs = new NotificationChainImpl();
 	public Resource myMetaModel;
-	static int user;
+	public static int user;
 	static Experience newXp = new Experience();
 	static Experience oldXp = new Experience();
 	static double factor = 0.0;
@@ -1425,7 +1425,7 @@ public class QLearning {
 
 	}
 
-	static void copyFile(File from, File to) throws IOException {
+	public static void copyFile(File from, File to) throws IOException {
 		Files.copy(from.toPath(), to.toPath());
 	}
 
@@ -1450,7 +1450,7 @@ public class QLearning {
 		}
 	}
 
-	static void createTags(int user) {
+	public static void createTags(int user) {
 		switch (user) {
 		//ECMFA paper preferences
 		case 0:
@@ -1532,101 +1532,4 @@ public class QLearning {
 			
 		}
 	}
-
-	public static void main(String[] args) throws IllegalAccessException, IllegalArgumentException,
-			InvocationTargetException, IOException, NoSuchMethodException, SecurityException {
-
-		QLearning ql = new QLearning();
-		user = 0;
-		ql.createTags(user);
-		// if there is previous knowledge
-		loadKnowledge();
-		long startTimeT = System.currentTimeMillis();
-		long endTimeT = 0;
-		String root = "././mutants/";
-		String root2 = "././fixed/";
-		File folder = new File(root);
-		File[] listOfFiles = folder.listFiles();
-		for (int i = 0; i < listOfFiles.length; i++) {
-			// invert mutant order
-			// for (int i = listOfFiles.length-1; i >=0; i--) {
-
-			File dest = new File(root2 + listOfFiles[i].getName());
-			long startTime = System.currentTimeMillis();
-			long endTime = 0;
-			System.out.println("----------------------------------------------------------------------");
-			System.out.println("----------------------------------------------------------------------");
-			System.out.println("STARTING WITH MODEL - " + i + ": " + listOfFiles[i].getName());
-
-			// Copy original file
-			copyFile(listOfFiles[i], dest);
-
-			ql.uri = URI.createFileURI(dest.getAbsolutePath());
-			ql.resourceSet.getResourceFactoryRegistry().getExtensionToFactoryMap().put("ecore",
-					new EcoreResourceFactoryImpl());
-			ql.myMetaModel = ql.resourceSet.getResource(ql.uri, true);
-
-			Resource auxModel = ql.resourceSet.createResource(ql.uri);
-			auxModel.getContents().addAll(EcoreUtil.copyAll(ql.myMetaModel.getContents()));
-
-			EPackage epa = (EPackage) auxModel.getContents().get(0);
-			System.out.println("Num. Classes: " + epa.getEClassifiers().size());
-
-			int enums = 0;
-			int refs = 0;
-			int attribs = 0;
-			int ops = 0;
-			int datatypes = 0;
-
-			for (int j = 0; j < epa.getEClassifiers().size(); j++) {
-				if (epa.getEClassifiers().get(j).getClass() == EEnumImpl.class) {
-					enums++;
-				} else {
-					if (epa.getEClassifiers().get(j).getClass() == EDataTypeImpl.class) {
-						datatypes++;
-					} else {
-						refs = refs + ((EClassImpl) epa.getEClassifiers().get(j)).getEAllReferences().size();
-						attribs = attribs + ((EClassImpl) epa.getEClassifiers().get(j)).getEAllAttributes().size();
-						ops = ops + ((EClassImpl) epa.getEClassifiers().get(j)).getEAllOperations().size();
-					}
-				}
-			}
-
-			System.out.println("Num. References: " + refs);
-			System.out.println("Num. Attributes: " + attribs);
-			System.out.println("Num. Operations: " + ops);
-			System.out.println("Num. Enums: " + enums);
-			System.out.println("Num. DataTypes: " + datatypes);
-			System.out.println("----------------------------------------------------------------------");
-			System.out.println("----------------------------------------------------------------------");
-
-			List<Error> errors = ql.errorsExtractor(auxModel);
-			ql.nuQueue = errors;
-
-			for (int y = 0; y < ql.nuQueue.size(); y++) {
-				ql.originalCodes.add(ql.nuQueue.get(y).getCode());
-			}
-			System.out.println("INITIAL ERRORS:");
-			System.out.println(errors.toString());
-			System.out.println("Size= " + errors.size());
-			System.out.println();
-
-			ql.actionsExtractor(errors);
-
-			ql.modelFixer(auxModel);
-
-			endTime = System.currentTimeMillis();
-			long timeneeded = (endTime - startTime);
-			System.out.println("TOTAL TIME: " + timeneeded);
-		}
-
-		System.out.println("COMPLETELY FINISHED!!!!!!");
-		endTimeT = System.currentTimeMillis();
-		long timeneededT = (endTimeT - startTimeT);
-		System.out.println("TOTAL EXECUTION TIME: " + timeneededT);
-		save(ql.getNewXp(), "././knowledge.properties");
-		System.exit(0);
-
-	}
-
 }
